Adapt to (u)int32_t int -> long switch-over.
GCC defined (u)int32_t as long on Xtensa, but Espressif patched this
to int up until ESP-IDF 4.4. Since IDF 5.0, (u)int32_t are long.

+++ software/src/modules/meters_evse_v2/meter_evse_v2.cpp
@@ -70,7 +70,7 @@ void MeterEVSEV2::update_from_evse_v2_all_data(EVSEV2MeterData *meter_data)
         state->get("type")->updateUint(meter_type);
 
         MeterValueID ids[METER_ALL_VALUES_RESETTABLE_COUNT];
-        uint32_t id_count = METER_ALL_VALUES_RESETTABLE_COUNT;
+        size_t id_count = METER_ALL_VALUES_RESETTABLE_COUNT;
         sdm_helper_parse_values(meter_type, all_values, &id_count, ids, value_packing_cache);
         value_count = id_count;
         meters.declare_value_ids(slot, ids, id_count);
