+++ software/src/modules/meters_rs485_bricklet/meter_rs485_bricklet.cpp
@@ -56,7 +56,7 @@ void MeterRS485Bricklet::changeMeterType(size_t supported_meter_idx)
     state->get("type")->updateUint(meter_type);
 
     MeterValueID ids[METER_ALL_VALUES_RESETTABLE_COUNT];
-    uint32_t id_count = METER_ALL_VALUES_RESETTABLE_COUNT;
+    size_t id_count = METER_ALL_VALUES_RESETTABLE_COUNT;
     sdm_helper_get_value_ids(meter_type, ids, &id_count);
     meters.declare_value_ids(slot, ids, id_count);
